New Polish Translation
I added new translation for the mod.
I have noticed that RDKRACZ is contributing to the mod as well. If he could check on the translation, I would appreciate it greatly.
While I think my translation is alright, I do think that having extra pair of eyes to potentially catch some errors would be reassuring.

I want to point out that this translation is not meant to be literal and direct in all cases, but it still tries to be faithful and keep original meaning.
Sometimes however, I had to completely change one sentence with another not related to it if out of context, but still fitting for specific situation. (I'm referring specifically to advancements titles - those were... a bit more on creative end of things)